requirements.txt
openai==1.84.0
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0
colorama==0.4.6
markdown==3.5.2 